fix: address review findings on #2287 — relocate scopeDegraded tests to smoke layer
The 3 scopeDegraded test cases in auth.unit.test.ts built a real Hono app
and exercised the middleware via app.request() — the smoke-layer contract
per CLAUDE.md's Test conventions table, not the unit layer. Moved them into
auth.smoke.test.ts's existing "scope resolver" describe block and extended
its shared makeAuthApp's /whoami handler to also expose scopeDegraded,
reusing the existing fakeAdminTokenService/fakeAgentTokenService helpers
instead of duplicating them in a second file.
